+++ nucleoid/CHANGELOG.md
@@ -1,3 +1,9 @@
+## 0.1.4
+- FileData new mime-type: gif
+- FileData default mime-type: application/octet-stream
+- Features ImageResize extends FileData, for file types: jpg, png, gif
+- Change styleList type: List<StyleCSS>
+
 ## 0.1.3
 - Add CSS property: opacity, user-select